Point static image notebook link at short version for RTD
The pointer in our documentation for the static image rendering example
now points to a short version of the real notebook with just one image.
Hopefully this is short enough to render successfully on RTD.  This
is just an interim fix but it will be good enough if it works.